Remove cmake
It's just too much of a pain to maintain cmake, it was constantly behind
and broken, and converting it over to the new gensio modules was going
to be hard.  And it didn't work with the new C++ swig stuff and go.

My main reason for having it was to use it under Visual Studio, which
turned out to not be very useful.  So I have no reason to keep it.

Signed-off-by: Corey Minyard <cminyard@mvista.com>