pty,stdio: Change the way euid and uid are used
The way that was being use was to set the user of a new process was to
set the euid to the new uid and the code would handle it.  It would set
everything up and then set the uid to the euid.  That's ok for Linux,
which creates the pty as the real uid, but it's not ok for BSD, which
doesn't.

So, change the code to set the real userid to the one you want it to be
and leave the euid alone.  This will give you full privileges while
setting everything up.  Then grantpt() will set the pty's slave owner to
the real uid.  The code will then set the effective (and saved) uid to
the real one before the exec() call.  This will work under BSD.

Signed-off-by: Corey Minyard <cminyard@mvista.com>