sound: Fix problems with windows sound output
When starting transmission, make sure that two buffers are available to
avoid stutter at the beginning.

When transmitting, if the end is reached and there is a partial buffer,
send it.  This required allowing partial buffers to be sent, which
required some modificiations to how data buffer lengths were done.

Fix header position size computations on conversion transmission.

Signed-off-by: Corey Minyard <cminyard@mvista.com>