afskmdm: Fix some transmission timing issues
When finishing a transmission, allow the transmission to complete and go
idle after all the data has been transmitted. Don't go idle after all
the data has been put into the queue, as it still needs to be sent.

Allow multiple message to be sent in one transmission, up to 256 bytes
plus the last message, to be more fair.

Fix the ranomization to actually work.

Signed-off-by: Corey Minyard <cminyard@mvista.com>