win: Fix PTYs started on other users under Windows
Windows PTYs have an issue where if the PTY is created at a higher
integrity level than the program using the console, that program cannot
do "wrong way verb" operations like WriteConsoleInput().  Unfortunately,
MSYS2 does WriteConsoleInput, and this breaks things.

Solve this by creating a helper program.  When a PTY is created and
an impersonation token is present, don't create the PTY in the main
program, instead create a helper program that will create the PTY and do
the I/O to the console.  The main program will talk to the helper
program over stdio and a special control pipe it creates.

Signed-off-by: Corey Minyard <cminyard@mvista.com>