telnet: Fix issues with sync handling
According to my memory of the telnet spec, you are supposed to get a
byte of data with exception data, then later comes the command byte and
the data mark to mark the end of the sync.

However, the sync handling in some cases didn't work as expected.  In
some cases, the command byte was coming before the exception data, and
then the actual data mark was coming as the exception data.  Rework the
sync handling to handle this.

Signed-off-by: Corey Minyard <cminyard@mvista.com>