Add a parameter handling infrastructure
The current parameter handling in gensio doesn't give a way to report
meaningful parsing errors to the user.  Add a new infrastructure for
gensios to use that allows error reporting back through the event
handler.

Have the old parsing functions call the new ones to avoid duplication
of function.

Move gensio_enum_val to gensio_types.h so it's available to the class
include.

Signed-off-by: Corey Minyard <cminyard@mvista.com>