cm109gpio: Rework the soundcard to HID device matching
It turns out using the container id doesn't work.  The container id
appears to be for the USB hub device, not the USB composite device.

Rework this to go up the parent chain looking for a USB composite
device for both the sound device and the HID device, then compare that
device id for a match.

Also, the way this works is to get the device and then use
SetupDiGetEnumDeviceInterfaces() to enumerate the interfaces in the
device.  The code I stole before was just completely wrong.

Signed-off-by: Corey Minyard <cminyard@mvista.com>