afskmdm: Add crc and checkax25 options
Add an option so the CRC processing can be disabled in the afskmdm
layer, primarily for testing AX.25 CRC processing.

Add an option to check that the packet is a valid AX.25 packet (address
and length are valid).  This is for test processing without an AX.25
layer to make sure packets are good.

Signed-off-by: Corey Minyard <minyard@acm.org>