afskmdm: Fix a possible overflow on a long message
Messages past the maximum length were just kind of kept but no new data
was added after the max size and the total size was just kept.  However,
this would cause an overrun when duplicating a wmsg, since the length
was not checked there.

There is no reason to keep the message around once it exceeds the
length, so just discard it.

Also, since the length value can no longer exceed the maximum value,
we can drop that check before processing a message.

Signed-off-by: Corey Minyard <minyard@acm.org>