osops: Fix a problem in Unix with groups
The code that handling setting a new user would properly set the
supplementary groups properly if a user was in supplementary groups.
However, it would not drop the groups if the new user didn't belong to
any supplementary groups, leaving the previous user's groups.

Add a call to drop the groups properly in that case.

Signed-off-by: Corey Minyard <minyard@acm.org>