base: Stop the timer before starting it on (dis)connect
On a connect and disconnect, if the filter function requested a timer to
be started, it would start it without stopping it.  If the timer was
running, this would be ignored, and you couldn't change the timeout.

Change this to stop the timer and then start the timer so it will
properly reset the timer.

Signed-off-by: Corey Minyard <minyard@acm.org>