ssl,certauth: Add a connection timeout handler
When connecting, start a timer that times the operation.  If the
connection doesn't complete in the given timeout, abort the connection.
This should avoid some denial of service issues.

As part of this, the try_connect() handling in the base code had to be
redone.  It was screwed up, anyway, and hacked special for certauth.
Clean that up and make adapt certauth to be more reasonable.

Signed-off-by: Corey Minyard <minyard@acm.org>