mdns: Fix the windows code
It turns out the Windows interface for service registration didn't work
much like anything I expected.  The docs are terrible, per normal
Microsoft policy I guess, and my guesses were mostly wrong.

Anyway, this fixes it.  There is a comment in the code that discusses
how it work, see that for details.

Signed-off-by: Corey Minyard <cminyard@mvista.com>