mdns: More Windows fixes
I finally figured out how everything works, I think.

For watches, the browser callback gets called with the full current
state of the IP addresses for the watch.  So adjust for that.  That also
means there is no need for a timer, so delete that code.

Rework the service callback.  It looks like you get one and only one
callback on the service registration, so handle appropriately.

Signed-off-by: Corey Minyard <cminyard@mvista.com>