c++:swig:go:tests: Fix the mdns tests
Several issues brought up by dnssd...

The service now needs to be waited for removal, or the oshandler count
will be wrong.

You cannot use globs on types with dnssd (and Windows), move the glob
test to the name.

Only wake the watch once, don't want a bunch of pending wakeups to mess
other things up.

Signed-off-by: Corey Minyard <minyard@acm.org>