osops: Add an option to use openpty
On macos posix_openpt returns an fd that you can't do much with, you
can't do O_NONBLOCK or fcntl() on it.  After some testing, you have to
actually open the client to set non-blocking, but just use openpty()
on that platform.

This required some unfortunate changes to the os handler because openpty
returns the child device as well :(.

Signed-off-by: Corey Minyard <minyard@acm.org>