Fix an issue with openssl on macos
On macos, on ARM, openssl uses sigsetmask() in it's startup code to
check some things.  If this happens after threads are started, sigsetmask
on macos sets the signal mask for all threads, not just the calling
thread, screwing up the sigmasks.

Add some code to detect this issue, and if detected, link ssl with the
main program so the startup code runs at startup time, not when the
openssl library is loaded by the ssl or certauth gensios.

Signed-off-by: Corey Minyard <minyard@acm.org>