mdns: Fix a number of issues, mostly on Windows
Fix issues with the typestr and domainstr on DNSSD and Windows, remove
the leading '=' and don't allow globs or regex on those, since it won't
work.

txt was not being processed correctly on Windows.

Add pcre regular expressions for Windows.

Discourage the use of globs, they don't work on Windows.

Update docs for building on Windows and MacOS.

Signed-off-by: Corey Minyard <cminyard@mvista.com>