conacc: Fix shutdown while in open issue
If an open call from conacc is not going to return in a timely manner,
like it's a server relpkt or mux, you can't stop the operation quickly
because the connacc code in that case waits for the open to complete
before handling the shutdown (or disable).  Modify the code to close
the connection in the shutdown or disable if the open is in progress,
and then finish the shutdown when the close completes.

For instance, if you started gensiot with:

ools/gensiot -a 'conacc,relpkt(mode=server),msgdelim,serialdev,/dev/ttyPipeA0,115200n81,local'

Doing a ^C on it would take a long time, basically until gensio gave up
on a timeout.  With this change, the close happens immediately.

Signed-off-by: Corey Minyard <minyard@acm.org>