serialdev: Split out GENSIO_EVENT_SER_xxxSTATE to add a mask event
The GENSIO_EVENT_SER_MODEMSTATE and GENSIO_EVENT_SER_LINESTATE
constants were used to report on the server that the client has
requested a mask change.  On the client it was used to report a
modemstate event.

This caused a problem on ser2net when a serial port was on the accepter
side.  When the modemstate event came in, ser2net didn't know that it
was a modemstate change report, it was assuming that it was a modemstate
mask request.  Then it would do a sergensio_modemstate() operation to
pass on the mask change, which would cause the serialdev gensio to
send another modemstate event.

This was kind of based on how the rfc2217 spec was written, but ser2net
needs to know if it was a mask change request or a state report.

So this change creates a GENSIO_EVENT_SER_xxxSTATE_MASK event for each
of this.  This way ser2net can tell if it's a mask request or a report
and do the appropriate thing.

Signed-off-by: Corey Minyard <minyard@acm.org>