telnet: Process telnet commands one at a time
The telnet processing code was processing all received command in a big
block.  If read was not enabled, you only want to process to a certain
point and then stop processing, but commands were getting injected
after.

Change the command processing to return after each command so the read
processing can check if the init is done and turn off read.

Signed-off-by: Corey Minyard <minyard@acm.org>