serialdev: Fix some issues with hangup-when-done
This option really wasn't working.  On Unix, it would restore the
original termios and close, so the setting would have no effect.  It
wasn't implemented on Windows.

Modify the way it works so the default is -1, which means don't modify
the value.  Whatever the default is is what will happen.  If you set it
to true or false on Unix, it will modify the value after the device is
closed, because there's no other way to do it.

Windows is now fixed so that if you set this it will drop DTR and RTS on
close if set.  Otherwise it has no effect.

Also, the default is renamed to match the parameter name.

Signed-off-by: Corey Minyard <minyard@acm.org>