Get rid of trying to guess the gensio type
If a gensio didn't match, it would try to guess the gensio type to be
compatible with the old version of ser2net.  That old version is long
not supported, and old configuration file support has been removed, so
get rid of this, as it results in confusing error messages.

Signed-off-by: Corey Minyard <minyard@acm.org>