Rework the modemstate handling once again.
Leave the handling of GENSIO_CONTROL_SER_MODEMSTATE and
GENSIO_CONTROL_SER_LINESTATE as they are.  They are kind of broken, but
they do sort of work.

They are broken because the set mask commands are suppose to result in
responses being sent.

Add GENSIO_CONTROL_SER_SEND_MODEMSTATE and
GENSIO_CONTROL_SER_SEND_LINESTATE for sending the modem and line state
reports.

Also add GENSIO_ACONTROL_SER_SET_MODEMSTATE_MASK and
GENSIO_ACONTROL_SER_SET_LINESTATE_MASK for the commands and responses
for the client and server.

Also, flush is a command that is supposed to get a response.  Modify it
to do so.

Signed-off-by: Corey Minyard <minyard@acm.org>