selector: Rework broken pselect handling
There were a couple of issues:

A signal may have been sent before this process was queued on a
wait queue, so we may have the signal pending without having been
queued or had the time set to zero.  Detect that and set things up
properly.

There really need to be some barriers around the time setting to
make sure things get set properly.

To do this, we needed to split out pselect into our own version
and use select() instead.  So some time handling had to be redone.

Signed-off-by: Corey Minyard <minyard@acm.org>