m4:openssl: Fix issues detecting ssl includes
I installed the msys version of the openssl development libraries
and this stoped working.  There was code for adding /ucrt64 and /mingw64
in the case of mingw, but it added those directories last instead of
first and it added both no matter what.

Check the MSYSTEM variable to find the right include and add that first
in the search list.  Also add mingw32 first in the search list, too.

Signed-off-by: Corey Minyard <minyard@acm.org>