osops: Store the right value for current break on Unix
The tty ioctls for Unix do not provide a way to fetch the current value
of break (TIOCSBRK, TIOCCBRK).  So the gensio library assumes it's off
at startup (well, turns it off unless disablebreak is off) and stores
the current value.

Unfortunately, it was storing the wrong value, and it would always
return "on" after it had been set on or off once.

Store the right value.

Signed-off-by: Corey Minyard <corey@minyard.net>