gltsshd: Rework program handling
Pass command from the remote as a set of argvs instead of a big command
string.

It doesn't work on Windows; windows quoting is such a mess.  Just pass
the argv and let the Windows-specific code hanadle it.

On Unix, it works fine either way, so don't go through all the message
conversion to a string and back again.

Signed-off-by: Corey Minyard <corey@minyard.net>