ax25: Delay the final channel close until the base closes
The final channel close on ax25 was happening when the channel closed,
but the base was still in the process of closing.  If a user used the
final channel close to know when they were done, and never called
any gensio waiters, it would leave the gensios under ax25 open.

Delay the close of the final channel until after the base code has
finished closing.

Signed-off-by: Corey Minyard <corey@minyard.net>