Fix a local address binding issue
If ipv6 only is set on a socket (you use ipv6 explicitly or implicitly
by doing nothing) it must be done before an address is bound to the
socket.  This means it must be done in socket setup before the bind,
not in the connect where it was done.
.
This required an API change to the os funcs because it needs
the address to know to set ipv6.

Signed-off-by: Corey Minyard <minyard@acm.org>